+++ src/Instance/HoldingsList/Holding/HoldingAccordion.js
@@ -17,6 +17,11 @@ import {
 } from '../../../hooks';
 
 import HoldingButtonsGroup from './HoldingButtonsGroup';
+import useHoldingItemsQuery from '../../../hooks/useHoldingItemsQuery';
+import {
+  useHoldingsAccordionState,
+  useLocationsQuery,
+} from '../../../hooks';
 
 const HoldingAccordion = ({
   children,
@@ -28,13 +33,16 @@ const HoldingAccordion = ({
   tenantId,
   isViewHoldingsDisabled,
   isAddItemDisabled,
+  instanceId,
+  pathToAccordionsState,
 }) => {
   const searchParams = {
     limit: 0,
     offset: 0,
   };
 
-  const [open, setOpen] = useState(false);
+  const pathToAccordion = [...pathToAccordionsState, holding?.id];
+  const [open, setOpen] = useHoldingsAccordionState({ instanceId, pathToAccordion });
   const [openFirstTime, setOpenFirstTime] = useState(false);
   const { totalRecords, isFetching } = useHoldingItemsQuery(holding.id, { searchParams, key: 'itemCount', tenantId });
   const { data: locations } = useLocationsQuery({ tenantId });
@@ -121,12 +129,16 @@ HoldingAccordion.propTypes = {
   holding: PropTypes.object.isRequired,
   onViewHolding: PropTypes.func.isRequired,
   onAddItem: PropTypes.func.isRequired,
+  instanceId: PropTypes.string.isRequired,
   holdings: PropTypes.arrayOf(PropTypes.object),
   withMoveDropdown: PropTypes.bool,
   children: PropTypes.object,
   tenantId: PropTypes.string,
   isViewHoldingsDisabled: PropTypes.bool,
   isAddItemDisabled: PropTypes.bool,
+  pathToAccordionsState: PropTypes.arrayOf(PropTypes.string),
 };
+
+HoldingAccordion.defaultProps = { pathToAccordionsState: [] };
 
 export default HoldingAccordion;