+++ src/Instance/InstanceDetails/InstanceDetails.js
@@ -128,6 +128,8 @@ const InstanceDetails = forwardRef(({
     );
   }
 
+  const isConsortialHoldingsVisible = instance?.shared || isInstanceShadowCopy(instance?.source);
+
   const renderPaneTitle = () => {
     const isInstanceShared = Boolean(isShared || isInstanceShadowCopy(instance?.source));
 
@@ -205,8 +207,8 @@ const InstanceDetails = forwardRef(({
               />
             )}
 
-            {instance?.shared && (
-              <ConsortialHoldings
+            {isConsortialHoldingsVisible && (
+               <ConsortialHoldings
                 instance={instance}
                 userTenantPermissions={userTenantPermissions}
               />