+++ src/Instance/InstanceDetails/MemberTenantHoldings/MemberTenantHoldings.js
@@ -17,6 +17,7 @@ import { MoveItemsContext } from '../../MoveItemsContext';
 
 import { useInstanceHoldingsQuery } from '../../../providers';
 import { hasMemberTenantPermission } from '../../../utils';
+import { useHoldingsAccordionState } from '../../../hooks';
 
 import css from './MemberTenantHoldings.css';
 
@@ -29,7 +30,13 @@ const MemberTenantHoldings = ({
     name,
     id,
   } = memberTenant;
+  const instanceId = instance?.id;
   const stripes = useStripes();
+
+  const pathToAccordion = ['consortialHoldings', id, '_state'];
+  const pathToHoldingsAccordion = ['consortialHoldings', id];
+  const [isMemberTenantAccOpen, setMemberTenantAccOpen] = useHoldingsAccordionState({ instanceId, pathToAccordion });
+
   const { holdingsRecords, isLoading } = useInstanceHoldingsQuery(instance?.id, { tenantId: id });
   const isUserInCentralTenant = checkIfUserInCentralTenant(stripes);
 
@@ -45,7 +52,8 @@ const MemberTenantHoldings = ({
       className={css.memberTenantHoldings}
       id={`${name}-holdings`}
       label={name}
-      closedByDefault
+      open={isMemberTenantAccOpen}
+      onToggle={() => setMemberTenantAccOpen(prevValue => !prevValue)}
     >
       <div className={css.memberTenantHoldings}>
         {isLoading
@@ -61,6 +69,7 @@ const MemberTenantHoldings = ({
                 isViewHoldingsDisabled={!canViewHoldings}
                 isAddItemDisabled={!canCreateItem}
                 isBarcodeAsHotlink={canViewItems}
+                pathToAccordionsState={pathToHoldingsAccordion}
               />
             </MoveItemsContext>
           )}