+++ src/ViewInstance.js
@@ -919,6 +919,7 @@ class ViewInstance extends React.Component {
       },
     ];
     const isInstanceLoading = this.state.isLoading || !instance || isCentralTenantPermissionsLoading;
+    const keyInStorageToHoldingsAccsState = ['holdings'];
 
     return (
       <DataContext.Consumer>
@@ -947,6 +948,7 @@ class ViewInstance extends React.Component {
                         instance={instance}
                         draggable={this.state.isItemsMovement}
                         tenantId={okapi.tenant}
+                        pathToAccordionsState={keyInStorageToHoldingsAccsState}
                         droppable
                       />
                     </MoveItemsContext>