+++ src/hooks/index.js
@@ -3,6 +3,7 @@ export { default as useBrowseValidation } from './useBrowseValidation';
 export { default as useCallout } from './useCallout';
 export { default as useHoldingItemsQuery } from './useHoldingItemsQuery';
 export { default as useHoldingMutation } from './useHoldingMutation';
+export { default as useHoldingsAccordionState } from './useHoldingsAccordionState';
 export { default as useInstanceMutation } from './useInstanceMutation';
 export { default as useHoldingsQueryByHrids } from './useHoldingsQueryByHrids';
 export { default as useInventoryBrowse } from './useInventoryBrowse';